requirements.txt
amqp==2.4.1
aniso8601==3.0.0
ansible==2.7.10
ansible-runner==1.3.4
arrow==0.12.1
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.2.1
attrs==17.4.0
Babel==2.6.0
bcrypt==3.1.6
billiard==3.5.0.4
bitmath==1.3.1.2
blinker==1.4
celery==4.2.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
coloredlogs==9.0
coverage==4.5.1
cryptography==2.5
Cython==0.28.1
defusedxml==0.5.0
dill==0.2.8.2
docopt==0.6.2
docutils==0.14
ecdsa==0.13
Flask==0.12.2
Flask-BabelEx==0.9.3
Flask-Celery==2.4.3
Flask-Cors==3.0.3
Flask-JWT-Extended==3.7.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Redis==0.3.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SocketIO==3.0.2
flask-swagger-ui==3.6.0
Flask-WTF==0.14.2
flower==0.9.2
future==0.17.1
get==1.0.3
gevent==1.3.6
gevent-websocket==0.10.1
gitdb2==2.0.5
GitPython==2.1.8
greenlet==0.4.15
gunicorn==19.9.0
haproxyadmin==0.2.2
httplib2==0.11.3
humanfriendly==4.17
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jose==1.0.0
kombu==4.3.0
lazy-object-proxy==1.3.1
ldap3==2.4.1
lockfile==0.12.2
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.1.0
Naked==0.1.31
npyscreen==4.10.5
oauth2==1.9.0.post1
oauthlib==2.1.0
paramiko==2.4.2
passlib==1.7.1
pexpect==4.7.0
pluggy==0.6.0
post==1.0.2
prompt-toolkit==1.0.15
psutil==5.5.0
psycopg2==2.7.6.1
ptyprocess==0.6.0
public==2.0.1
py==1.5.3
pyasn1==0.4.2
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.4.2
pylint==2.1.1
PyNaCl==1.3.0
pytest==3.5.0
pytest-cov==2.5.1
pytest-flask==0.10.0
python-daemon==2.2.3
python-dateutil==2.7.2
python-dotenv==0.8.2
python-engineio==3.3.1
python-http-client==3.1.0
python-jose==3.0.1
python-memcached==1.59
python-socketio==3.1.2
python3-openid==3.1.0
pytz==2018.3
PyYAML==3.13
query-string==1.0.2
redis==2.10.6
request==1.0.2
requests==2.19.1
requests-oauthlib==1.0.0
rsa==4.0
scp==0.10.2
shellescape==3.4.1
six==1.11.0
smmap2==2.0.5
social-auth-app-flask==1.0.0
social-auth-core==2.0.0
speaklater==1.3
tabulate==0.8.2
tornado==5.1.1
typed-ast==1.1.0
uniqid==0.0.1
urllib3==1.23
vine==1.1.4
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
